moved tests and formatted
